TFG/QuantumSoup/TFG_LuciaPereira_QuantumSoup-main/QuantumSoup/requirements.txt
Django>=5.1.1
djangorestframework
qiskit
requests
matplotlib
pylatexenc